Fixed alot of the css bugs.
Fixed some scaling issue bugs. Still need to fix the issue where when the page size along the x gets to small, and the scroll bar shows up, it causes the images to align to much to the left and on the right side causes issues with the "body" element outsizing the rest of the elements on the page. Besides that it works appropriately for all resizes and forms and maintains a certain size limit in all directions.