Updated some of the views, and added some more javascript
Updated the views so me and sannan are merged and everything is working.
Seperated the javascript for clouds and fall guys and have them spawn less on views which are not the homepage.
Fixed some of the layout stuff. Alot more to still do.
Fixed some of the footer stuff and some styles.