More For UI Overhaul WIP
Fixed some of the layout stuff. Added dynamic colour changing svg's for the side bar.
Made the side bar collapseable, and hide the text when it is collapsed.
Made the side bar show on each page
Add some styling to the side bar.
Added a inner scrollbar to the side bar to scroll when the page is to small to fit it all.
Made it so that the main content of the page cant go and shrink the nav bar to small to prevent seeing everything.