UI Overhaul still in progress
Got the ability to keep the sidebar and accordion the correct state when going through pages.
Fixed a issue when clicking on the sidebar link and switching pages, the page would reload, and set it to the original design for the sidebar, causing a small glitch, this is now fixed and is loading correctly with no weird visual things, fragmeneted the code across all the files, adding in a better maincontent area which will help keep things more centered. Changed the colours of the text and keep the page that is currently loaded highlighted in the sidebar so we can visually see where we are.